Scaffolded optional support for output caching
In addition to new error handling, OnTopic 5.2.0 also supports optional output caching. We don't want to enable this for GoldSim, as it interferes with the server-side cookie handling of the cookie acceptance notification. But I've registered placeholders for it, in case we want to revisit this in the future.

(Warning: This requires moving the `CookieNoticeViewComponent` to work off of client-side cookies, which introduces a potential Cumulative Layout Shift—and, thus, isn't currently in place.)