Preserve original license option when toggling research license type
When selecting the "GoldSim Research" product, the only license type that will be available is the "Leased Standalone (1 year)" option. If a different license type was selected prior to selecting "GoldSim Research", this will now store that locally, and then reselect it when "GoldSim Research" is unchosen (either by choosing "GoldSim" or "Select one…").