Bug: Allow license type to be selected prior to product
Due to the logic for reverting the license type to the previously selected value (6fd4d15), a bug occurred where the license would be reset if it was selected prior to the product. This is fixed by only resetting the value if it's already been set (which happens when "GoldSim Research" is selected).