Instead of hiding options, treat license options as read-only
Previously, we hid all options not related to the "Research GoldSim" product. Now, instead of hiding these, I am mimicking a read-only state on the dropdown box using the newly introduced `select.readonly` class (840e890). This looks and acts like a disabled dropdown list, but will still submit the value to the server.