Replaced explicit font declarations with `Base/fonts`
The `_fonts.scss` imports the same Google fonts—but in a single payload. It also imports the same Font Awesome font—but also two additional fonts which may be needed for some styles, and is thus more reliable. This ensures that the same exact fonts available to `Base/_base.scss` are available to the `BaseStyle.scss`.