Added download icons to `BaseStyle`
The `Base/_icons.scss` file will prefix any link that points to common download formats (such as `zip`, `docx`, `pdf`, `gsm`, &c.) with the appropriate brand icon for that format, thus making it clear that the link is to a download, and of what type. This will be useful in CkEditor for inline links, but may also be useful for the Help site, for PDFs in particular.