Expanded target of `_links` to include MadCap Flare windows
The `_links` stylesheet doesn't apply to the entire `body` because we don't want it to be applied to header, navigation, and footer elements. Instead, it's applied to the `main` block. This isn't used on OnTopic Editor, and thus the use of the `.editor` selector. Similarly, for the MadCap Flare documentation, we want this applied to the `div` with the `role` set to `main`, as well as to the breadcrumbs (via the `.breadcrumbs` proxy).