Introduce two-column definition lists
By default, definition lists are displayed with the definition (`dd`) on a new line, indented. This changes them to operate off of a two-column format. This was originally intended for the print version of the documentation, but MadCap Flare's PDF converter doesn't recognize this. Nevertheless, it's a preferable baseline for `dl`'s than the default behavior.