Updated `CallsToAction` view to conditionally anchor to `Announcement`
Based on the new `HasAnnouncement` bit in `ViewData`, conditionally anchor to either the `SiteFooter` or `SiteAnnouncement`. As part of this, I renamed the `FooterAnnouncement` to `SiteAnnouncement` since it is also used in the header on the homepage—even if it's not anchored to the `CallsToAction` view component on the homepage.