Differentiated Page-Level `NavigationTopicViewModel`
A custom `NavigationTopicViewModel` is used for the `PageLevelNavigationViewComponent`. For simplicity, however, and to ensure cache results were centralized, it had been used across all menus, not just the page-level navigation.

This is unnecessary, however, as the page-level navigation child topics are actually excluded from the other navigation topics, since they're meant to be covered by the page-level navigation. As a result, not only does this clutter up the data model for most of the menu view components, but the logical naming convention—`NavigationTopicViewModel`, for consistency with `OnTopic.ViewModels`—creates a conflict with the global `using` of `OnTopic.ViewModels`.

To remedy this, as part of moving the view component related view models to `GoldSim.Web.Models.Components`, I've renamed `NavigationTopicViewModel` to `PageLevelNavigationTopicViewModel`, thus avoiding the naming conflict, and created a new caching container for it in the `GoldSimActivator.cs`, such that the other menus can go back to using the standard `NavigationTopicViewModel`.