Refactored view components to use `WebPath` over `Key`
One benefit to previously using the now-`PageLevelNavigationTopicViewModel` for other navigation view components, and not just the page-level navigation, is that it exposed the `Key` property, which the standard `NavigationTopicViewModel` doesn't.

With the `PageLevelNavigationTopicViewModel` now _only_ being used by the `PageLevelNavigationViewComponent` (7ddb8a6a), that is no longer available to the other view components. As such, these have been modified to instead rely on `WebPath`. For the main Menu's `Title` fallback, this requires parsing the `WebPath` to get the last topic key from it.