Introduced a new `AnnouncementViewModel`
Previously, I had reused the existing `AssociatedTopicViewModel` (d4cbaa5f). This was acceptable, as it meets the current requirements, and doesn't add much more overhead. It's also not intuitive, however, and prevents us from extending the view model in the future if we need to add e.g. `Location` or `LinkLabel`.

To prevent that, I've added a new `AnnouncementViewModel`. For now, it's just `Label` and `Url`.

Since it's so simple, I'm manually mapping it, instead of establishing a dependency on the `ITopicMappingService`, and requiring it to be done via reflection, which is slower.