Introduced a new `CallsToActionViewModel`
Previously, I had used the `ViewData` to relay the `HasAnnouncement` bit (6e4fa264, 9ae63397). This was acceptable, as it meets the current requirements, and doesn't add much more overhead. It's also not intuitive, however, since it's not part of the view model. By migrating to a custom `CallsToActionViewModel`, the `HasAnnouncement` is a first-class citizen built into the data model, and we're also able to extend the `CallsToAction` view model further in the future now that we have this scaffolded.