Allow PDF icons on links with a `#`
The icon selectors expect the file name to be at the end of the `href` value. That's a generally valid assumption—and a faster selector. But for PDFs, that won't necessarily be true, as sections of a PDF can be linked using a `#`. To accommodate this, the selector for PDFs—and PDFs only—is updated to search for `.pdf` anywhere within the `href` attribute.