Fixed bug in Course's `IsPreview` implementation
Previously, I introduced a new `IsPreview` feature for the Courses so that a course that isn't completed can be hidden from non-authenticated users (4fa2ac80). In doing so, I assume that the `CurrentTopic` is valid—but since the `OnActionExecuting()` method of `CoursesController` happens prior to the inherited `TopicController`'s `OnActionExecuting()`, this isn't guaranteed, as the `CurrentTopic` hasn't been validated. As a result, a 500 error is thrown due to a `NullReferenceException` instead of a 404 error due to a missing `Topic`. This is easily fixed with a null-conditional operator in the call to `CurrentTopic`.