Replaced Advanced Topics list with open field
Instead of asking participants about a predetermined list of advanced topics that they are interested in, this now simply provides a text area for them to provide an open list. The goal is not to bias or restrict the topics based on GoldSim's assumptions. As part of this, I've marked the `AdvancedTrainingTopicsSelection` as `Obsolete`—but am keeping it around for now, just in case we decide to rehydrate it later.