Added support for MathJax to OnTopic
This includes the latest MathJax plugin for CKEditor 2.x, as well as the 2.7.4 version of MathJax itself. Note that while the frontend of the website uses the latest MathJax 3.x (6f3e9bb4), the CKEditor plugin only supports MathJax 2.x, and thus these are not in sync. For the level of equations that GoldSim is working with, this isn't expected to cause an issue, since the TeX markup will be consistent between the two, and no extensions are being taken advantage of. Should it become an issue, however, this can be changed in the future.