Established a new `TopicSearchViewModel`
This will store both the `Query` and the `Results` of the search. I've placed this in a new `GoldSim.Web.Models.Controllers` namespace, which can be used for view models associated with controllers, similar to the `GoldSim.Web.Models.Components` namespace, which is used for view models associated with components. I've added this new namespace to the `_ViewImports.cshtml` file, so all views are aware of it.