Fixed bug with Include Training selection
It's unclear how this worked previously, but the value of the `IncludeTraining` field wasn't compatible with the `bool` in the Binding Model. This is fixed by changing the value to a Boolean value. Alternatively, the backing field could have been changed to a string, at which point this would have been a valid value.