Updated `FollowUpPage` to be a Topic Reference
In a prior version of OnTopic, topic references were elevated from standard attributes to first-class citizens. This change was reflected in the Content Type, but not in the data model. As a result, GoldSim was still pulling the legacy values for `FollowUpPage` from the Topic's attributes and not its references. This update fixes that issue by mapping `FollowUpPage` to a `CoreTopicViewModel`, which represents a topic reference.