Reintroduced `Layout` definitions in nested templates
Some of the `ContentList` templates are embedded in other templates, which don't have a `Layout` defined. For example, `WhitePaperList` includes `LinkedList` as a partial. For this to work, the "parent" template must set `Layout` to empty. But this overrides the `Layout` set in `_ViewStart`. As such, when I (re)introduced the `/Views`-specific `_ViewStart`, and removed all of the `Layout` definitions within the templates (61227952), this introduced a bug where any pages embedding `ContentList` pages were throwing errors due to a missing Layout. In practice, this only affected the White Paper List page, since the other templates that do this don't appear to be actively used at this point. (Or they called `Accordion`, which still maintained its `Layout` definition for some reason, thus avoiding this bug.)