Replaced `IsPreview` with existing `IsPrivateBranch` to prevent crawling
In a previous update, I introduced a new `IsPreview` bit for branches (4fa2ac80). This is a novel attribute and relies on custom logic to enforce. I've updated this to instead use the existing `IsPrivateBranch` under-the-hood. This is useful because the underlying `SitemapController` from OnTopic hides all Sitemap entries from a Topic flagged accordingly (see OnTopic/OnTopic-Library#72). In practice, because the `CourseController` didn't allow anonymous users to view these pages, they would never get indexed—but they would throw errors in the Google Search Console due to the HTTP 401 response. As such, this not only hides these courses from the Sitemap, but also unifies the nomenclature with an established OnTopic attribute. This corresponds with an update to the `IsPreview` attribute in GoldSim's OnTopic database.