Hide Preview Courses from Course List
Courses marked as being in a preview state—previously with the `IsPreview` attribute, now changed to the `IsPrivateBranch` attribute (0d01ef30)—cannot be viewed by anonymous users. However, they are still listed in the `CourseListViewComponent` despite that. To hide them, they must also be marked as hidden from navigation, using the `IsHidden` attribute. To avoid remembering to take this extra step, I'm updating the `UnitListViewComponent`—which the `CourseListViewComponent` inherits from—to exclude courses marked as `IsPrivateBranch` from being listed.