Updated paths to reflect available files
Several of the paths in the `Scripts` section of the `gulpFile` weren't present. This was exposed during a trial upgrade to Gulp 5.0.0. While I encountered other errors that I wasn't able to resolve at this point—and will need to reevaluate later—this at least exposed the fact that these paths were incorrect, which apparently the current Gulp 4.0.2 version simply overlooks.

In the case of Greensock and ScrollMagic, this isn't expected to matter, as we really should only be exposing the minified versions.

As far as what happened to Headroom, it's unclear—but it hasn't been available in our `node_modules` folder for some time, nor is it a direct dependency, nor is it called directly from any of our JavaScript code, so hopefully that means it isn't necessary.