Conditionally add replace UI to view
If a search has already been conducted, a replace input and replace button are added. The replace button is separate so that a modification to the search string doesn't accidentally result in erasing all search results.

Later, we'll want to add a warning popup to this to confirm submission. For now, I'm adding the `alert` class to the "Replace" button so it's clear that it's doing something invasive.