Move classes, members to `internal`
As these needn't be called externally, there's currently no benefit to having them be `public`—and if that changes, we should evaluate it on a per case basis. The two exceptions are interface implementations, which must remain public, and some instances of default constructors, as required for e.g. data binding.

This resolves a warning, `ca1515`.