Provided concrete type when known
Instead of casting a type as its interface, rely on the concrete type when it's known ahead of time. If, in the future, we need this to support multiple implementations we can move back to the interface declaration.

Previously, this was preferred because the constructor injection is looking for the interface, so that makes the relationship clear.

This resolves CA1859.