Returned View Components to `public`
Previously, as per code analysis recommendations, most classes and members in the project were moved from `public` to `internal` (90ab905d), since they aren't needed externally. But view components are required to be `public` in order to be visible to the ASP.NET Core runtime. As a result, I've migrated all view components back to `public`.