Returned types to `public`
This extends the changes made to controllers (fea78fdd), view components (adaee788), and view models (b2ecf8b0) to the various areas.

Previously, as per code analysis recommendations, most classes and members in the project were moved from `public` to `internal` (90ab905d), since they aren't needed externally. But controllers and view components are required to be `public` in order to be visible to the ASP.NET Core runtime, as are view models to be visible to OnTopic. As a result, I've migrated all controllers, view components, and view models back to `public`.