Fix business logic for `scope`
In my original implementation of the business logic for scope (ca7562de), my approach was mistaken, and has no basis for working correctly. Here, I fix that by instead handling the scope in the main controller action, by relying on `GetByUniqueKey()` to find the relevant topic. This requires first converting the web path to a unique key.