Applied `form-field` classes
Added `form-field` classes so that the forms display consistently with other forms on the site. While I was at it, I moved to a more modern approach for `<label>`, wrapping the `<input>` in the `<label>`, instead of relying on the `for` attribute to link them together. This allowed me to get rid of the `id` attribute on each of the `<input>` fields.