Use Foundation input label for the new scope field
This helps differentiate the `Scope` field from the `Search` and `Replace` fields, and makes it really clear what scope means. As part of this, I needed to pull in the Foundation Forms library, which we weren't previously using. I also changed the default value from `/Web/` to `Web`, which makes more sense when also using the label.