Added support to toggle regular expressions
Regular expressions can cause confusion for day-to-day operations, and especially for those not familiar with them. To address this, I've added a new checkbox to conditionally enable regular expressions, an `UseRegEx` backing property in `TopicSearchViewModel` to capture that setting, and then conditional logic in the `TopicSearchController` to respond to that setting.