Don't show previews for non-RegEx searches
When searching with regular expressions enabled, both the query matches and the replacement results can vary on a per match basis, and thus it's critical to provide a preview. When regular expressions aren't being used, however, these values will always be the same as the original query and replace expressions, and thus don't provide any value. As such, when regular expressions are disabled, I'm instead just summarizing the number of matches. As part of this, I introduced a new `pluralize()` helper method to aid in conditionally adding an "s" to plural labels.