Add in optional columns, but leave them blank
Optionally, each entitlement can be assigned to an Account ID or explicitly assigned an Expiration Date. These columns are looked for by ordinal, so we want them to be represented before the informational columns (i.e., the ones which aren't needed by FLiP). This also gives GoldSim the option of assigning these if e.g. the customer already has an account or they want to set a different expiration date for a particular request for some reason.