Updated rasterized logos with new logo
Ideally, we should be using the new `Logo.svg` file. For backward compatibility with legacy applications which may be using the PNG formats, however, I'm updating the two PNG files—`Logo.png` and `Logo.Small.png`—with the new logo. As part of this, I'm maintaining the legacy height of the image instead of cropping it to fit the new type face, as this will maximize backward compatibility.