Conditionally set action for global search
When a user searches a term on the global search bar while on a course, the search results will be displayed on the search template under that specific course. That way, the interface remains the same when doing the search, and the context of the course is maintained.

This also works in concert with the new `scope` option, which uses that as a cue to limit the search to that course (a9cfc95a).