Remove empty type declaration bodies
As of C# 14, we no longer need to define an empty body on a type if it doesn't implement any logic of its own. This is a common scenario for view models, which need to have their own names to be mapped to unique Razor templates, but may not change the underlying implementation of the content type view model that they derive from.