Fix path to Administration script
There isn't a `LicenseAdministration` script. Presumably, this was looking for the `Administration` script.

That said, it's not immediately clear to me that this script is even necessary on the edit page, so that may be a mistake. It won't hurt anything, and the file should already be cached, but it's running some jQuery queries that might waste some processes.