Made view models bound to views public
View models bound to views can be marked `internal`. Unfortunately, however, while the ASP.NET Core runtime accepts this, ReSharper / Rider throw warnings about "inconsistent accessibility". Since it's really not a big deal if these are public, I'm returning them to be so, which is the default to .NET tends to use anyway. Ones not bound to Razor templates will remain `internal`.