Suppress false positive regarding unreachable code
ReSharper / Rider think this is a false positive because they seem to assume that the project is using nullability annotations, which it isn't. Otherwise, they'd realize that `GetInvoice()` can, indeed, return null, and therefore this block is reachable. Since we need this check, we'll manually suppress these warnings.

(In other cases, we were able to do the checks, or suppress them with the null-forgiving operator, as per 25efc361, a170145c., and 736dcd29.)