Update Code Analyzers to 10.0
This corresponds with the update to .NET 10.0 (1841a1cf). This doesn't affect the runtime behavior. But this is expected to expose a lot of suggestions related to not only new capabilities of .NET 10 and C# 14, but also smarter recommendations built into the upgraded analyzers. Those will be handled in separate updates.